asset changes: moved upload attachment to the sidebar, added a little structure to the show asset action, made it so that tools can enabled/disable the ability to add attachments.
I also commented out the code that allows you to create a page from
the attachment form. I can add it back if people want, but I would
like to talk about it a little... i thought the previous discussion
was that we would try to keep the idea of asset pages and attachments
separated, but i could be wrong. Now, if you set a title, it changes
the filename.



git-svn-id: http://code.autistici.org/svn/crabgrass/trunk@428 4da624d7-9686-4b4d-bf4a-96ca807925bb